Split .hpp file
